patient-dash/backend/requirements.txt
django
djangorestframework
django-cors-headers
requests
twilio
Pillow
boto3
django-storages
cloudinary
cryptography
